+++ netcall/green/service.py
@@ -21,13 +21,10 @@ Authors:
 # Imports
 #-----------------------------------------------------------------------------
 
-import exceptions
-from types import GeneratorType
-
 import zmq
 
-from ..base  import RPCServiceBase
-from ..utils import logger, get_zmq_classes, detect_green_env, get_green_tools
+from ..base_service import RPCServiceBase
+from ..utils        import logger, get_zmq_classes, detect_green_env, get_green_tools
 
 
 #-----------------------------------------------------------------------------
@@ -64,119 +61,17 @@ class GreenRPCService(RPCServiceBase):
 
         super(GreenRPCService, self).__init__(**kwargs)
 
-        _, _, _, _, self.Queue = get_green_tools(env=self.green_env)
-
         self.greenlet = None
-        self.yield_send_queues = {}  # {<req_id> : <Queue>}
         # Can also use collections.deque, append() and popleft() being thread safe
     #}
     def _create_socket(self):  #{
         super(GreenRPCService, self)._create_socket()
         self.socket = self.context.socket(zmq.ROUTER)
     #}
-    def _handle_request(self, msg_list):  #{
-        """Handle an incoming request.
-
-        The request is received as a multipart message:
-
-        [<id>..<id>, b'|', req_id, proc_name, <ser_args>, <ser_kwargs>, <ignore>]
-
-        First, the service sends back a notification that the message was
-        indeed received:
-
-        [<id>..<id>, b'|', req_id, b'ACK',  service_id]
-
-        Next, the actual reply depends on if the call was successful or not:
-
-        [<id>..<id>, b'|', req_id, b'OK',    <serialized result>]
-        [<id>..<id>, b'|', req_id, b'YIELD', <serialized result>]*
-        [<id>..<id>, b'|', req_id, b'FAIL',  <JSON dict of ename, evalue>]
-
-        Here the (ename, evalue, traceback) are utf-8 encoded unicode.
-
-        In case of a YIELD reply, the client can send a YIELD_SEND, YIELD_THROW or
-        YIELD_CLOSE messages with the same req_id as in the first message sent.
-        The first YIELD reply will contain no result to signal the client it is a
-        yield-generator. The first message sent by the client to a yield-generator
-        must be a YIELD_SEND with None as argument.
-
-        [<id>..<id>, b'|', req_id, 'YIELD_SEND',  <serialized sent value>]
-        [<id>..<id>, b'|', req_id, 'YIELD_THROW', <serialized ename, evalue>]
-        [<id>..<id>, b'|', req_id, 'YIELD_CLOSE', <no args & kwargs>]
-
-        The service will first send an ACK message. Then, it will send a YIELD
-        reply whenever ready, or a FAIL reply in case an exception is raised.
-
-        Termination of the yield-generator happens by throwing an exception.
-        Normal termination raises a StopIterator. Termination by YIELD_CLOSE can
-        raises a GeneratorExit or a StopIteration depending on the implementation
-        of the yield-generator. Any other exception raised will also terminate
-        the yield-generator.
-        """
-        req = self._parse_request(msg_list)
-        if req is None:
-            return
-        self._send_ack(req)
-
-        ignore = req['ignore']
-
-        try:
-            # raise any parsing errors here
-            if req['error']:
-                raise req['error']
-
-            if req['proc'] in ['YIELD_SEND', 'YIELD_THROW', 'YIELD_CLOSE']:
-                if req['req_id'] not in self.yield_send_queues:
-                    raise ValueError('req_id does not refer to a known generator')
-
-                self.yield_send_queues[req['req_id']].put((req['proc'], req['args']))
-                return
-            else:
-                # call procedure
-                res = req['proc'](*req['args'], **req['kwargs'])
-        except Exception:
-            not ignore and self._send_fail(req)
-        else:
-            if ignore:
-                return
-
-            if isinstance(res, GeneratorType):
-                self._handle_yield(req, res)
-            else:
-                self._send_ok(req, res)
-    #}
-    def _handle_yield(self, req, res):  #{
-        req_id = req['req_id']
-        logger.debug('Adding reference to yield %s', req_id)
-        input_queue = self.yield_send_queues[req_id] = self.Queue(1)
-
-        self._send_yield(req, None)
-        gene = res
-
-        try:
-            while True:
-                proc, args = input_queue.get() # Will get stuck when shutdown()
-
-                if proc == 'YIELD_SEND':
-                    res = gene.send(args)
-                    self._send_yield(req, res)
-
-                elif proc == 'YIELD_THROW':
-                    ex_class = getattr(exceptions, args[0], Exception)
-                    eargs = args[:2]
-                    eargs[0] = ex_class
-                    res = gene.throw(*eargs)
-                    self._send_yield(req, res)
-
-                else:
-                    gene.close()
-                    self._send_ok(req, None)
-                    break
-        except:
-            self._send_fail(req)
-        finally:
-            logger.debug('Removing reference to yield %s', req_id)
-            del self.yield_send_queues[req_id]
+    def _get_tools(self):  #{
+        "Returns a tuple (Queue, Empty)"
+        _, _, _, _, Queue, Empty = get_green_tools(env=self.green_env)
+        return Queue, Empty
     #}
     def start(self):  #{
         """ Start the RPC service (non-blocking).
@@ -190,13 +85,15 @@ class GreenRPCService(RPCServiceBase):
         spawn = get_green_tools(env=self.green_env)[0]
 
         def receive_reply():
-            while True:
+            self.running = True
+            while self.running:
                 try:
                     request = self.socket.recv_multipart()
                 except Exception, e:
                     logger.warning(e)
                     break
                 spawn(self._handle_request, request)
+            self.running = False
             logger.debug('receive_reply exited')
 
         self.greenlet = spawn(receive_reply)
