+++ netcall/threading/client.py
@@ -31,9 +31,9 @@ except ImportError:
 
 import zmq
 
-from ..base   import RPCClientBase
-from ..utils  import get_zmq_classes, ThreadPool, logger
-from ..errors import RPCTimeoutError
+from ..base_client import RPCClientBase
+from ..utils       import get_zmq_classes, ThreadPool, logger
+from ..errors      import RPCTimeoutError
 
 
 #-----------------------------------------------------------------------------
@@ -222,10 +222,10 @@ class ThreadingRPCClient(RPCClientBase):  #{
                 
                 if msg_type == b'YIELD':
                     futures_get_fn = futures.get
-                    future_init_fn = _ReturnOrYieldFuture.init_as_yield
+                    future_init_fn = self._ReturnOrYieldFuture.init_as_yield
                 else: # For OK and FAIL
                     futures_get_fn = futures.pop
-                    future_init_fn = _ReturnOrYieldFuture.init_as_return
+                    future_init_fn = self._ReturnOrYieldFuture.init_as_return
                 
                 future = futures_get_fn(req_id, None)
                 if future is None:
@@ -246,6 +246,10 @@ class ThreadingRPCClient(RPCClientBase):  #{
         req_sub.close(0)
 
         logger.debug('io_thread exited')
+    #}
+    def _get_tools(self):  #{
+        "Returns a tuple (Event, Queue, Future, TimeoutError)"
+        return Event, Queue, Future, TimeoutError
     #}
     def call(self, proc_name, args=[], kwargs={}, ignore=False, timeout=None):  #{
         """
@@ -284,6 +288,7 @@ class ThreadingRPCClient(RPCClientBase):  #{
         if timeout and timeout > 0:
             def _abort_request():
                 result = self._futures.pop(req_id, None)
+                not result.is_init() and result.init_as_return()
                 if result is not None:
                     tout_msg  = "Request %s timed out after %s sec" % (req_id, timeout)
                     logger.debug(tout_msg)
@@ -293,7 +298,7 @@ class ThreadingRPCClient(RPCClientBase):  #{
         else:
             timer = None
 
-        future = _ReturnOrYieldFuture(self, req_id)
+        future = self._ReturnOrYieldFuture(self, req_id)
         self._futures[req_id] = future
         #logger.debug('waiting for future=%r' % future)
         try:
@@ -324,64 +329,4 @@ class ThreadingRPCClient(RPCClientBase):  #{
             self.pool.stop()
             self.pool.join()
     #}
-#}
-
-class _ReturnOrYieldFuture(object):  #{
-
-    def __init__(self, client, req_id):
-        self.is_initialized = Event()
-        self.return_or_except = Future()
-        self.yield_queue = Queue(1)
-        self.client = client
-        self.req_id = req_id
-
-    def init_as_return(self):
-        assert(not self.is_init())
-        self.is_yield = False
-        self.is_initialized.set()
-
-    def init_as_yield(self):
-        assert(not self.is_init())
-        self.is_yield = True
-        self.is_initialized.set()
-
-    def is_init(self):
-        return self.is_initialized.is_set()
-
-    def set_result(self, obj):
-        assert(self.is_init())
-        if self.is_yield:
-            self.yield_queue.put(obj)
-        else:
-            self.return_or_except.set_result(obj)
-
-    def set_exception(self, ex):
-        assert(self.is_init())
-        self.return_or_except.set_exception(ex)
-        if self.is_yield:
-            self.yield_queue.put(None)
-
-    def _try_except(self):
-        try:
-            self.return_or_except.result(timeout=0)
-        except TimeoutError:
-            pass
-
-    def result(self):
-        self.is_initialized.wait()
-
-        if not self.is_yield:
-            return self.return_or_except.result()
-
-        else:
-            self._try_except()
-            self.yield_queue.get()
-
-            def recv_yielder():
-                while True:
-                    obj = self.yield_queue.get()
-                    self._try_except()
-                    yield None, obj
-
-            return self.client._yielder(recv_yielder(), self.req_id)
 #}
